+++ django_browserid/tests/test_views.py
@@ -8,7 +8,7 @@ from django.test.client import RequestFactory
 from mock import patch
 
 from django_browserid import views
-from django_browserid.tests import mock_browserid
+from django_browserid.tests import mock_browserid, patch_settings
 
 
 factory = RequestFactory()
@@ -123,3 +123,50 @@ def test_redirect_field_name():
                       redirect_field_name='my_redirect', **kwargs)
     assert response.status_code == 302
     assert response['Location'].endswith('/field_success')
+
+
+@patch_settings(DEBUG=True, SESSION_COOKIE_SECURE=True)
+@patch('django_browserid.views.log.debug')
+def test_sanity_session_cookie(debug):
+    # If DEBUG == True and SESSION_COOKIE_SECURE == True, log a debug message
+    # warning about it.
+    verify('post', assertion='asdf')
+    debug.called = True
+
+
+@patch_settings(DEBUG=True, MIDDLEWARE_CLASSES=['csp.middleware.CSPMiddleware'])
+@patch('django_browserid.views.log.debug')
+def test_sanity_csp(debug):
+    # If DEBUG == True, the django-csp middleware is present, and Persona isn't
+    # allowed by CSP, log a debug message warning about it.
+
+    # Test if allowed properly.
+    with patch_settings(CSP_DEFAULT_SRC=[],
+                        CSP_SCRIPT_SRC=['https://login.persona.org'],
+                        CSP_FRAME_SRC=['https://login.persona.org']):
+        verify('post', assertion='asdf')
+        debug.called = False
+    debug.reset_mock()
+
+    # Test fallback to default-src.
+    with patch_settings(CSP_DEFAULT_SRC=['https://login.persona.org'],
+                        CSP_SCRIPT_SRC=[],
+                        CSP_FRAME_SRC=[]):
+        verify('post', assertion='asdf')
+        debug.called = False
+    debug.reset_mock()
+
+    # Test incorrect csp.
+    with patch_settings(CSP_DEFAULT_SRC=[],
+                        CSP_SCRIPT_SRC=[],
+                        CSP_FRAME_SRC=[]):
+        verify('post', assertion='asdf')
+        debug.called = True
+    debug.reset_mock()
+
+    # Test partial incorrectness.
+    with patch_settings(CSP_DEFAULT_SRC=[],
+                        CSP_SCRIPT_SRC=['https://login.persona.org'],
+                        CSP_FRAME_SRC=[]):
+        verify('post', assertion='asdf')
+        debug.called = True